Use posts-public because of DDB…
DynamoDB filters after a limit is applied. So if we want to
exclude private and/or draft posts we end up with fewer
posts than requested. This commit copies posts to a
separate posts-public table that is used for lists. The
categories-posts table is also updated to only public posts.
If a request is made by a trusted client the original posts
table is queried.