Add Meetup/Eventbrite context retrieval
As I RSVP to events on Meetup/Eventbrite, it'd be useful to have the
Pipes that I've set up for retrieving metadata from them to be supported
in `fetch-context`.

These follow the format of Granary, but expose an `is..Url` helper so we
can simplify the logic when retrieving context.

Additionally, we only want to use them if properties are found for a
lookup, otherwise still fall back to Granary/OpenGraph.